requirements.txt
path==16.3.0
Pillow==9.0.1
qrcode==7.3.1